+++ client/components/astro-object-form/AstroObjectForm.tsx
@@ -146,6 +146,7 @@ const AstroObjectForm: React.FC<AstroObjectFormProps> = ({ disabled, initialData
             <div className={styles.sections}>
                 <div className={styles.inputSection}>
                     <Input
+                        size={'large'}
                         required={true}
                         disabled={disabled}
                         className={styles.formElement}
@@ -155,6 +156,7 @@ const AstroObjectForm: React.FC<AstroObjectFormProps> = ({ disabled, initialData
                     />
 
                     <Input
+                        size={'large'}
                         required={true}
                         disabled={disabled}
                         className={styles.formElement}
@@ -164,6 +166,7 @@ const AstroObjectForm: React.FC<AstroObjectFormProps> = ({ disabled, initialData
                     />
 
                     <Input
+                        size={'large'}
                         required={true}
                         disabled={disabled}
                         className={styles.formElement}
@@ -178,6 +181,7 @@ const AstroObjectForm: React.FC<AstroObjectFormProps> = ({ disabled, initialData
                     />
 
                     <Input
+                        size={'large'}
                         required={true}
                         disabled={disabled}
                         className={styles.formElement}
@@ -193,6 +197,7 @@ const AstroObjectForm: React.FC<AstroObjectFormProps> = ({ disabled, initialData
                     />
 
                     <Input
+                        size={'large'}
                         required={true}
                         disabled={disabled}
                         className={styles.formElement}